Add test resolve from parent pom
