+++ uberfire-m2repo-editor/uberfire-m2repo-editor-backend/src/test/resources/org/guvnor/m2repo/backend/server/helpers/gav-pom-without-group.xml
@@ -1,0 +1,16 @@
+<project xmlns="http://maven.apache.org/POM/4.0.0"
+         xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance"
+         xsi:schemaLocation="http://maven.apache.org/POM/4.0.0
+                             http://maven.apache.org/maven-v4_0_0.xsd">
+  <modelVersion>4.0.0</modelVersion>
+
+  <parent>
+    <artifactId>parent-gav-pom</artifactId>
+    <groupId>org.guvnor.test</groupId>
+    <version>1.0</version>
+  </parent>
+
+  <artifactId>testExistingParentGavInPom</artifactId>
+  <version>1.1.1</version>
+
+</project>
