+++ uberfire-m2repo-editor/uberfire-m2repo-editor-backend/src/test/resources/org/guvnor/m2repo/backend/server/parent-gav-pom.xml
@@ -1,0 +1,10 @@
+<?xml version="1.0" encoding="UTF-8"?>
+<project xmlns="http://maven.apache.org/POM/4.0.0" xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance"
+         xsi:schemaLocation="http://maven.apache.org/POM/4.0.0 http://maven.apache.org/xsd/maven-4.0.0.xsd">
+  <modelVersion>4.0.0</modelVersion>
+  <groupId>org.guvnor.test</groupId>
+  <artifactId>parent-gav-pom</artifactId>
+  <version>1.0</version>
+  <name>parent-gav-pom</name>
+  <packaging>pom</packaging>
+</project>
